Backend: Move `Stringifier` to the backend
