+++ loki/backend/__init__.py
@@ -8,10 +8,11 @@
 Backend classes that convert Loki IR into output code in various languages.
 """
 
-from loki.backend.fgen import * # noqa
 from loki.backend.cgen import * # noqa
 from loki.backend.cppgen import * # noqa
 from loki.backend.cudagen import * # noqa
-from loki.backend.pygen import * # noqa
-from loki.backend.dacegen import * # noqa
 from loki.backend.cufgen import * # noqa
+from loki.backend.dacegen import * # noqa
+from loki.backend.fgen import * # noqa
+from loki.backend.pygen import * # noqa
+from loki.backend.pprint import * # noqa
