+++ loki/backend/cgen.py
@@ -10,16 +10,14 @@ from pymbolic.mapper.stringifier import (
     PREC_UNARY, PREC_LOGICAL_OR, PREC_LOGICAL_AND, PREC_NONE, PREC_CALL
 )
 
+from loki.backend.pprint import Stringifier
+
+from loki.expression import (
+    symbols as sym, LokiStringifyMapper, Array, symbolic_op, Literal
+)
+from loki.ir import Import, FindNodes, FindVariables, FindRealLiterals
 from loki.logging import warning
 from loki.tools import as_tuple
-from loki.ir import (
-        Import, Stringifier, FindNodes,
-        FindVariables, FindRealLiterals
-)
-from loki.expression import (
-        LokiStringifyMapper, Array, symbolic_op, Literal,
-        symbols as sym
-)
 from loki.types import BasicType, SymbolAttributes, DerivedType
 
 __all__ = ['cgen', 'CCodegen', 'CCodeMapper', 'IntrinsicTypeC']
